requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pandas
pytest
coverage
numpy
httpx
memoized-property
gcsfs
termcolor
uvicorn
streamlit
python-dotenv
pillow
azure-cognitiveservices-vision-computervision
